chore(database): adds database service class
